refactor it a bit
