+++ src/oraku-influx2pg/src/oraku_influx2pg/version.py
@@ -1,0 +1,7 @@
+from importlib.metadata import version, PackageNotFoundError
+
+try:
+    # cannot derive distribution name from import name (__package__) unfortunately
+    __version__ = version("aare-oraku-influx2pg")
+except PackageNotFoundError:
+    __version__ = "0.0.0"
